WIP: refactor secret handling
